temp commit

+++ src/resource_mac.h
@@ -25,8 +25,8 @@ struct ResourceMacEntry {
 	uint16_t id;
 	uint16_t nameOffset;
 	uint32_t dataOffset;
-	uint32_t compressedSize;
-	uint32_t size;
+	uint16_t compressedSize;
+	uint16_t size;
 	uint8_t type;
 	char name[kResourceMacEntryNameLength];
 };
